Another try to fix FSDP

+++ mappo/agents/causal_full_agent.py
@@ -25,7 +25,7 @@ class CausalFullAgent(LlamaFullAgent):
         model_dtype = torch.float16 if self.device.startswith("cuda") else torch.float32
         self.base_model = AutoModelForCausalLM.from_pretrained(
             model_name,
-            torch_dtype=model_dtype,
+            dtype=model_dtype,
         )
         self.base_model.to("cpu")
