+++ mappo/agents/llama_full_agent.py
@@ -42,7 +42,7 @@ class LlamaFullAgent:
         model_dtype = torch.float16 if self.device.startswith("cuda") else torch.float32
         self.base_model = LlamaForCausalLM.from_pretrained(
             model_name,
-            torch_dtype=model_dtype,
+            dtype=model_dtype,
         )
         self.base_model.to("cpu")
